requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
pypdf==5.3.1
PyMuPDF==1.26.4
python-docx==1.1.2
pytest==8.4.1
